requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==2025.6
decorator==5.2.1
distlib==0.3.8
Django==5.1.1
django-crispy-forms==2.4
et_xmlfile==2.0.0
filelock==3.15.4
gunicorn==23.0.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
mysqlclient==2.2.7
narwhals==1.31.0
numpy==2.3.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.3.0
pipenv==2024.4.1
platformdirs==4.2.2
plotly==6.0.1
proglog==0.1.12
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
requests==2.32.5
setuptools==72.2.0
six==1.17.0
sqlparse==0.5.1
tqdm==4.67.1
tzdata==2024.1
urllib3==2.5.0
virtualenv==20.26.3
watchdog==6.0.0
wheel==0.44.0
whitenoise==6.7.0
